refactors filters/aggs #402, #265
