+++ lib/facets/create-aggregations.js
@@ -2,9 +2,6 @@
 * Create the facets for each categories
 */
 const aggregationAll = require('./aggs-all');
-const aggregationPeople = require('./aggs-people');
-const aggregationObjects = require('./aggs-objects');
-const aggregationDocuments = require('./aggs-documents');
 
 module.exports = function (queryParams) {
   var aggs = {};
@@ -14,15 +11,15 @@ module.exports = function (queryParams) {
   }
 
   if (queryParams.type === 'people') {
-    aggs.people = aggregationPeople(queryParams);
+    aggs.people = aggregationAll(queryParams);
   }
 
   if (queryParams.type === 'objects') {
-    aggs.objects = aggregationObjects(queryParams);
+    aggs.objects = aggregationAll(queryParams);
   }
 
   if (queryParams.type === 'documents') {
-    aggs.documents = aggregationDocuments(queryParams);
+    aggs.documents = aggregationAll(queryParams);
   }
   return aggs;
 };
