+++ lib/search.js
@@ -1,6 +1,4 @@
-const createFiltersPeople = require('./facets/create-filter-people');
-const createFiltersObjects = require('./facets/create-filter-objects');
-const createFiltersDocuments = require('./facets/create-filter-documents');
+const createFilters = require('./facets/create-filters');
 const createFiltersAll = require('./facets/create-filter-all');
 const aggregationTotalCategories = require('./facets/aggs-total-categories');
 const createAggregations = require('./facets/create-aggregations');
@@ -159,9 +157,9 @@ module.exports = (elastic, queryParams, next) => {
   }
 
   const filtersType = [{terms: {'type.base': ['agent', 'archive', 'object']}}];
-  const filtersPeople = createFiltersPeople(queryParams);
-  const filtersObjects = createFiltersObjects(queryParams);
-  const filterDocuments = createFiltersDocuments(queryParams);
+  const filtersPeople = createFilters(queryParams, 'agent');
+  const filtersObjects = createFilters(queryParams, 'object');
+  const filterDocuments = createFilters(queryParams, 'archive');
   const filters = {
     type: filtersType,
     people: filtersPeople,
