+++ lib/transforms/search-results-to-jsonapi.js
@@ -256,7 +256,7 @@ function createMeta (queryParams, results) {
   const total = (results.hits || {}).total || 0;
   const totalPages = Math.ceil(total / queryParams.pageSize);
   const countCategories = createCountCategories(queryParams.type, results);
-  return { total_pages: totalPages, count: {type: countCategories}, filters: createFacets[queryParams.type](results) };
+  return { total_pages: totalPages, count: {type: countCategories}, filters: createFacets(results) };
 }
 
 function createCountCategories (type, results) {
