+++ test/facets/aggs-people.test.js
@@ -3,7 +3,7 @@ const queryString = require('querystring');
 const createQueryParams = require('../../lib/query-params/query-params');
 const dir = __dirname.split('/')[__dirname.split('/').length - 1];
 const file = dir + __filename.replace(__dirname, '') + ' > ';
-const aggregationPeople = require('../../lib/facets/aggs-people');
+const aggregationPeople = require('../../lib/facets/aggs-all');
 
 test(file + 'The filters date are not included in the aggregation filters', (t) => {
   const query = queryString.parse('q=ada&filter%5Bdate%5Bfrom%5D%5D=wrongdate&filter%5Bdate%5Bto%5D%5D=wrongdate&page%5Bsize%5D=50');
